Fix scripts to adapt sql server 2022 breaking changes.
Note that with MSSQL Server 2022 new security features were added and it
is enforced by default to enable encryption to prevent man-in-the-middle
attacks. However since we use these scripts for local development, these
security features are irrelevant for us and does not worth to efforts in
setting up a certificate and such.
Due to the new changes, this also breaks some parts of our scripts where
it communicates with the database. Therefore we changed our approach to
enforce every query to trust the certificate and disable encryption.
That is also why I changed the way we create backup to use a stored
procedure instead, since the previous method no longer works using
dotnet modules.

+ Add colored messages.
+ Change the approach for creating database backups.
+ Change the approach of sending queries to the db from a centralized
  function.