Fix import script when generating rev-list to always return a list.
Note that in a situation where running the script with -generate_diffs
and only 1 migration script is running, the rev-list will return only a
single commit, which will be treated as a string instead of a list.
Therefore when indexing it with commits[0], in a case of a string, it
will return the first character instead of the first string as would be
expected.