Install automatically powershell sqlserver module.
Note that users often struggle with this step, since it either fails to
them due to missing admin rights or not following steps properly.
Therefore and to ease the initial process, we can install these modules
without admin permissions if we install it to the current user scope.
That way in the initial run it will install them automatically so we can
reduce the steps from the readme.